install: use symlinks instead of copies
This originally installed symbolic links but then was changed to support
a yet-to-be-implemented variable substitution scheme.  The links are
convenient, though, as they clearly indicate which files are managed
with this framework and which are not.

Install the files into an untracked subdirectory of the repository and
then link to those them.  This makes incremental installs safer (we
can't accidentally clobber files we don't manage.)